chore: renamed components, they don't use Menu anymore
